Added a new API function and made REPL::open_file a bit smarter.
I needed an API function that would configure a Prima::Edit widget
for me, but not create the new widget in a new tab for me. So, I
created REPL::endow_editor_widget. I also updated the open_file
function to be smarter about when it should create a new tab. If
the current tab's default widget is an editor object, it opens the
file into that editor; otherwise it creates a new tab and opens the
text into that file buffer.