Finally fixed the newline output bug.
After refactoring the InputHistory into its own module, the following
four lines of intput:

 print "Hello"
 1
 print "Hello\n"
 1

gave the following lines of output

 > print "Hello"
 Hello> 1
 > print "Hello\n"
 Hello
 > 1

The second line, in which the "this is what you gave me" printout
is tagged onto the end of Hello, is not what I wanted. This commit
finally fixes that by creating an new output handler for InputHistory
that checks the column number of the output widget and moves to the
next line if it's not in column zero.