Simplified CaptureStdOut.pm a bit
I came across Tk::Stdio in PAR::Packer and decided to compare it to what I
had in my current CaptureStdOut. I noticed that the Tk implementation used
tie() and untie() in a clean and symetric way. I wondered why I hadn't
used tie() and untie() in my stuff, and when I looked again at my code, I
realized that there was no reason not to use them.

I believe that the original design had hoped to play nicely with other
STDOUT/STDERR capturing mechanisms, but that is in the misty past, and it
certainly was not true of the implementation that preceeds this one. At
this point, Prima::CaptureStdOut only works well with multiple copies of
itself, and not with other capturing mechanisms.